+++ source/runner/_tests/fixtures/__DangerfileTypeScript.ts
@@ -1,3 +1,4 @@
+// @ts-ignore
 /* tslint-disable */
 
 // This doesn't exist in JS-world